% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordmeta_wrappers.R
\name{F_i}
\alias{F_i}
\title{Beta probability for ith order statistic}
\usage{
F_i(p, i, n)
}
\arguments{
\item{p}{p-value (scalar)}

\item{i}{rank (1..n)}

\item{n}{number of inputs}
}
\description{
Beta probability for ith order statistic
}
