% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACAT_wrappers.R
\name{magcat_pmn_file}
\alias{magcat_pmn_file}
\title{Path to built-in PMN pathway file}
\usage{
magcat_pmn_file(species = c("maize", "sorghum", "arabidopsis", "plant"))
}
\arguments{
\item{species}{Which built-in DB to use. For now:
\itemize{
\item{"maize"}   {corncyc\_pathways.20230103}
\item{"sorghum"} {sorghumbicolorcyc\_pathways.20230103}
\item{"arabidopsis"} {aracyc\_pathways.20230103}
\item{"plant"}  {plantcyc\_pathways.20230103}
}}
}
\description{
Path to built-in PMN pathway file
}
\keyword{internal}
