% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFisher_wrappers.R
\name{magcat_tpm_stat}
\alias{magcat_tpm_stat}
\title{Internal: truncated Fisher (TPM) statistic + approximate p-value}
\usage{
magcat_tpm_stat(p, ptrunc = 0.5)
}
\arguments{
\item{p}{numeric vector of p-values.}

\item{ptrunc}{scalar in (0, 1]; p-values < ptrunc are included.}
}
\value{
list(stat, nincl, nexcl, validp, ptrunc, pval)
}
\description{
Uses:
stat = -2 * sum(log p_i), over p_i < ptrunc
df   = 2 * nincl
If no p-values pass truncation (nincl = 0), returns stat = 0, pval = 1.
}
