% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_wrappers.R
\name{magma_annotate}
\alias{magma_annotate}
\title{Run MAGMA annotation}
\usage{
magma_annotate(snp_loc, gene_loc, out_prefix, window = NULL)
}
\arguments{
\item{snp_loc}{Path to SNP location file.}

\item{gene_loc}{Path to gene location file.}

\item{out_prefix}{Output prefix for MAGMA.}

\item{window}{Optional length-2 numeric vector (kb upstream, kb downstream).
For example, window = c(25, 25) gives "window=25,25".}
}
\description{
Run MAGMA annotation
}
