% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_competitive_geneset_wrappers.R
\name{magma_geneset_competitive}
\alias{magma_geneset_competitive}
\title{Run MAGMA competitive gene-set (pathway) analysis}
\usage{
magma_geneset_competitive(
  gene_results_raw,
  set_annot,
  out_prefix,
  out_dir = NULL
)
}
\description{
Run MAGMA competitive gene-set (pathway) analysis
}
