% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMNIBUS_ACAT_minP_wrappers.R
\name{omni_pathways}
\alias{omni_pathways}
\title{Omnibus pathway test (ACAT-O or minP) using MAGCAT \emph{prepared} objects}
\usage{
omni_pathways(
  gene_results,
  pathways = NULL,
  species = NULL,
  pmn_gene_col = NULL,
  gene_col = "GENE",
  p_col = "P",
  weight_col = NULL,
  ptrunc = 0.05,
  min_p = 1e-15,
  do_fix = TRUE,
  B_perm = 0L,
  perm_mode = c("resample_global", "mvn"),
  magma_genes_out = NULL,
  magma_cor_file = NULL,
  make_PD = TRUE,
  seed = NULL,
  omnibus = c("ACAT", "minP"),
  remove_singletons = TRUE,
  output = FALSE,
  out_dir = "magcat_omni"
)
}
\description{
Uses your wrapper pipeline for:
(1) analytic method p-values  (via *_prepare() + *_run_prepared())
(2) omni-level permutations   (resample_global or mvn)
}
\details{
Methods included (ONLY): ACAT, Fisher, soft-TFisher, minP, Stouffer
}
