% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_wrappers.R
\name{magma_gene}
\alias{magma_gene}
\title{Run MAGMA gene analysis on SNP p-values}
\usage{
magma_gene(
  bfile,
  gene_annot,
  pval_file,
  n_total,
  snp_col = "SNP",
  p_col = "P",
  out_prefix,
  gene_model = "snp-wise=top"
)
}
\arguments{
\item{bfile}{PLINK prefix for LD reference.}

\item{gene_annot}{MAGMA gene annotation file (.genes.annot).}

\item{pval_file}{SNP p-value file.}

\item{n_total}{Total sample size (N=... in MAGMA).}

\item{snp_col}{Column name in pval_file containing SNP IDs.}

\item{p_col}{Column name in pval_file containing p-values.}

\item{out_prefix}{Base output prefix. A suffix per gene model will be added.}

\item{gene_model}{Character vector of one or more MAGMA gene models,
e.g. "snp-wise=top", "snp-wise=mean", "multi".}
}
\description{
Run MAGMA gene analysis on SNP p-values
}
